+++ src/utils.cpp
@@ -2,7 +2,7 @@
 #include <string>
 #include <iostream>
 #include <fstream>
-bool is_close(double a, double b, double tolerance){
+bool is_close(float a, float b, float tolerance){
     return std::abs(a - b) < tolerance;
 }
